Ajustes na exibição do user

+++ resources/views/users/index.blade.php
@@ -30,25 +30,27 @@
    <th width="280px">Ação</th>
  </tr>
  @foreach ($data as $key => $user)
-  <tr>
-    <td>{{ ++$i }}</td>
-    <td>{{ $user->name }}</td>
-    <td>{{ $user->email }}</td>
-    <td>
-      @if(!empty($user->getRoleNames()))
-        @foreach($user->getRoleNames() as $v)
-           <label class="badge badge-success">{{ $v }}</label>
-        @endforeach
-      @endif
-    </td>
-    <td>
-       <a class="btn btn-info" href="{{ route('users.show',$user->id) }}">Visualizar</a>
-       <a class="btn btn-primary" href="{{ route('users.edit',$user->id) }}">Editar</a>
-        {!! Form::open(['method' => 'DELETE','route' => ['users.destroy', $user->id],'style'=>'display:inline']) !!}
-            {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
-        {!! Form::close() !!}
-    </td>
-  </tr>
+  @if($user->getRoleNames()[0] == 'Admin' )
+    <tr>
+        <td>{{ ++$i }}</td>
+        <td>{{ $user->name }}</td>
+        <td>{{ $user->email }}</td>
+        <td>
+          @if(!empty($user->getRoleNames()))
+            @foreach($user->getRoleNames() as $v)
+              <label class="badge badge-success">{{ $v }}</label>
+            @endforeach
+          @endif
+        </td>
+        <td>
+          <a class="btn btn-info" href="{{ route('users.show',$user->id) }}">Visualizar</a>
+          <a class="btn btn-primary" href="{{ route('users.edit',$user->id) }}">Editar</a>
+            {!! Form::open(['method' => 'DELETE','route' => ['users.destroy', $user->id],'style'=>'display:inline']) !!}
+                {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
+            {!! Form::close() !!}
+        </td>
+      </tr>
+    @endif
  @endforeach
 </table>
